MappingNormalizer: fix groupBy mappings dropping rename/enum/expression PM extends
Pre-fix, synthesizeSourceSpec gated the entire 'addDynaFunctionExtends'
step on '~groupBy is absent'. That step actually emitted four PM kinds
(rename, enum, expression-access, dyna), so under '~groupBy' all four
were skipped — leaving rename/enum/expression PMs without their
logical-named extends and breaking the body's terminal schema for any
aggregate mapping.

Split into two honestly-named functions:

  addRenameEnumExpressionExtends  — always runs (per-row, even under groupBy)
  addDynaScalarExtends            — only when no ~groupBy (under ~groupBy
                                    these become aggregate cols inside
                                    addMappingGroupBy)

addMappingGroupBy now picks the logical (PM property) name for a key
when a simple-rename PM aliases the physical groupBy column — the
rename extend has already added that name to the source schema, so the
groupBy(... ~[name] ...) Z⊆T subset check passes and the body terminal
matches the synth function's declared Relation<{props}> return type.

Schema-honesty side effects:

* TypeChecker emits an actionable hint on primitive-type schema
  mismatches pointing the user at the three honest fixes (change class
  type, change column type, or add an explicit conversion).

* 4 test files declared 'salary/total/price: Float[1]' against DECIMAL
  columns. These were genuine inconsistent-type bugs the strict schema
  check now catches; switched the class properties to Decimal[1] to
  match the storage (currency/money is the natural fit). JSON-source
  cases stay Float because get(@Float) extracts as Float.

* MultiFileRebuildTest: declare EnumerationMappings for Status/Side
  enums in both V1 and V2 mappings (canonical Pure shape; was relying
  on permissive enum coercion that we no longer emit silently).

The forwardRelationalRename hook in MappingResolver is the temporary
scaffolding that lets these inline rename extends layer into the class
StoreResolution. It will be removed in a following commit as part of
the larger Resolver rewrite.

+++ engine/src/test/java/com/gs/legend/test/JsonMappingIntegrationTest.java
@@ -180,7 +180,7 @@ class JsonMappingIntegrationTest {
                     Class model::Order {
                         id: Integer[1];
                         customerName: String[1];
-                        total: Float[1];
+                        total: Decimal[1];
                     }
                     """ + DATABASE_SCALAR + """
                     import model::*;
@@ -216,7 +216,7 @@ class JsonMappingIntegrationTest {
                         orderId: Integer[1];
                         productName: String[1];
                         quantity: Integer[1];
-                        price: Float[1];
+                        price: Decimal[1];
                     }
                     """ + DATABASE_SCALAR + """
                     import model::*;
